y2038: unify to uint64 GetTickCount64()
Addresses #602
Keep only uint64 GetTickCount64() as a source of ticks in
milliseconds. Update all calls to use this function and to
store values in uint64 datatypes.
GetTickCount64() will use internally clock_gettime() instead of
the deprecated gettimeofday().
Additionally, reduce calls by removing unused variables and too
consecutive calls that can be stored.

+++ src/EMSocket.h
@@ -60,10 +60,10 @@ public:
 	// CUpDownClient::TickDownloadAndMeasure.
 	void	WakeIfPaused();
 
-	virtual uint32	GetTimeOut() const;
-	virtual void	SetTimeOut(uint32 uTimeOut);
+	virtual uint64	GetTimeOut() const;
+	virtual void	SetTimeOut(uint64 uTimeOut);
 
-    uint32	GetLastCalledSend() { return lastCalledSend; }
+    uint64	GetLastCalledSend() { return lastCalledSend; }
 
     uint64	GetSentBytesCompleteFileSinceLastCallAndReset();
     uint64	GetSentBytesPartFileSinceLastCallAndReset();
@@ -150,9 +150,9 @@ private:
     bool m_currentPackageIsFromPartFile;
 
 	bool	m_bAccelerateUpload;
-	uint32	lastCalledSend;
-    uint32	lastSent;
-	uint32	lastFinishedStandard;
+	uint64	lastCalledSend;
+	uint64	lastSent;
+	uint64	lastFinishedStandard;
 
     uint32 m_actualPayloadSize;
     uint32 m_actualPayloadSizeSent;
